add images + clients

+++ src/AppBundle/Controller/UsersController.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Controller;
 
 use AppBundle\Entity\User;
+use AppBundle\Entity\UpdateAt;
 use AppBundle\Exception\ResourceValidationException;
 use AppBundle\Representation\Users;
 use FOS\RestBundle\Controller\Annotations as Rest;
@@ -12,6 +13,7 @@ use FOS\RestBundle\Controller\FOSRestController;
 use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\HttpFoundation\Response;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Cache;
+use Symfony\Component\HttpFoundation\Request;
 
 
 /**
@@ -45,11 +47,20 @@ class UsersController extends FOSRestController
      *     default="0",
      *     description="The pagination offset"
      * )
-     * @Rest\View()
+     * @Rest\View(StatusCode = 200)
      */
     public function listAction(ParamFetcherInterface $paramFetcher)
     {
+        $client=$this->get('security.token_storage')->getToken()->getUser();
+
+		$requestTime=Request::createFromGlobals()->headers->get('Last-Modified');
+		$dbTime=$this->getDoctrine()->getManager()->getRepository('AppBundle:UpdateAt')->findOneByTable("user")->getUpdatedAt();
+		if($requestTime==$dbTime){
+			return new Response($requestTime,Response::HTTP_OK);
+		}
+
         $pager = $this->getDoctrine()->getRepository('AppBundle:User')->search(
+            (string)$client,
             $paramFetcher->get('keyword'),
             $paramFetcher->get('order'),
             $paramFetcher->get('limit'),
@@ -65,12 +76,16 @@ class UsersController extends FOSRestController
      *     name = "app_user_show",
      *     requirements = {"id"="\d+"}
      * )
-     * @Rest\View
+     * @Rest\View(StatusCode = 200)
      * 
      * @Cache(lastModified="user.getUpdatedAt()")
      */
     public function showAction(User $user)
     {
+		if($this->get('security.token_storage')->getToken()->getUser()!=$user->getClient()){
+			throw $this->createAccessDeniedException('You cannot access this user !');
+		}
+		
         return $user;
     }
 
@@ -91,6 +106,7 @@ class UsersController extends FOSRestController
         }
 
         $em = $this->getDoctrine()->getManager();
+		$user->setClient($this->get('security.token_storage')->getToken()->getUser());
 
         $em->persist($user);
         $em->flush();
@@ -109,6 +125,10 @@ class UsersController extends FOSRestController
      */
     public function updateAction(User $user, User $newuser, ConstraintViolationList $violations)
     {
+		if($this->get('security.token_storage')->getToken()->getUser()!=$user->getClient()){
+			throw $this->createAccessDeniedException('You cannot update this user !');
+		}
+
         if (count($violations)) {
             $message = 'The JSON sent contains invalid data. Here are the errors you need to correct: ';
             foreach ($violations as $violation) {
@@ -119,10 +139,12 @@ class UsersController extends FOSRestController
         }
 
         $em=$this->getDoctrine()->getManager();
-        $user->setTitle($newuser->getTitle());
-        $user->setContent($newuser->getContent());
-        $user->setContent($newuser->getContent());
-        $user->setAuthor($newuser->getAuthor());
+        $user->setName($newuser->getName());
+        $user->setFirstName($newuser->getFirstName());
+        $user->setEmail($newuser->getEmail());
+        $user->setPassword($newuser->getPassword());
+        $user->setUpdateAt(new \DateTime());
+        $user->setClient($this->get('security.token_storage')->getToken()->getUser());
 
         $em->flush();
 
@@ -130,7 +152,7 @@ class UsersController extends FOSRestController
     }
 
     /**
-     * @Rest\View(StatusCode = 200)
+     * @Rest\View(StatusCode = 204)
      * @Rest\Delete(
      *     path = "/api/users/{id}",
      *     name = "app_user_delete",
@@ -139,6 +161,9 @@ class UsersController extends FOSRestController
      */
     public function deleteAction(User $user)
     {
+		if($this->get('security.token_storage')->getToken()->getUser()!=$user->getClient()){
+			throw $this->createAccessDeniedException('You cannot delete this user !');
+		}
         $em=$this->getDoctrine()->getManager();
         $em->remove($user);
         $em->flush();
